UPDATE: working version of lightning out, but putting on hold
many of the links are done with the navigator in mind.
even if we get lightning out to work, it won’t work in classic
because classic doesn’t understand the navigator OR the pages the tiles are set to
like /lightning/o/Account/home

so something would need to translate it back to classic….
and that isn’t the spirit of what we’re trying to do…

further work would be needed to make the tiles aware if they are in lightning out
and only show tiles that would - but they would ONLY work in classic if they did
because the URLs are so different.